qeuyn

+++ local/app/Http/Controllers/Admin/BannerController.php
@@ -20,14 +20,28 @@ class BannerController extends Controller
 		$banner->banner_head2 = $request->banner_head2;
         $banner->banner_position = $request->banner_position;
         $banner->banner_description = $request->banner_description;
-		
-		if( $request->has('banner_image') ){
-			foreach( explode(',',$banner->banner_image) as $image ){
-				File::delete('local/storage/app/public/home/image/'.$image,'local/storage/app/public/home/image/resized-'.$image);
-			}
+//        dd($request->all());
+        $video = $request->banner_image;
+//        dd($video);
+        $banner->banner_video = $request->banner_video;
+        if ($request->banner_video != null ) {
 
-			$banner->banner_image = saveImage([$request->banner_image],100,'public/home/image');
-		}
+            $filename = time() . '.' . $video->getClientOriginalExtension();
+            $banner->banner_image = $filename;
+            $path = public_path().'/home/image';
+            $video->move($path, $filename);
+        }
+        else{
+            if( $request->has('banner_image') ){
+                foreach( explode(',',$banner->banner_image) as $image ){
+                    File::delete('local/storage/app/public/home/image/'.$image,'local/storage/app/public/home/image/resized-'.$image);
+                }
+
+                $banner->banner_image = saveImage([$request->banner_image],100,'public/home/image');
+            }
+//            $banner->banner_image = saveImage([$request->banner_image],100,'public/home/image');
+        }
+
 		
 		$banner->save();
 		return back();
@@ -39,7 +53,19 @@ class BannerController extends Controller
 		$banner->banner_head2 = $request->banner_head2;
 		$banner->banner_description = $request->banner_description;
 		$banner->banner_position = $request->banner_position;
-		$banner->banner_image = saveImage([$request->banner_image],100,'public/home/image');
+        $banner->banner_video = $request->banner_video;
+        $video = $request->file('file');
+        dd($request->all());
+        if ($request->banner_video != null ) {
+
+            $filename = time() . '.' . $video->getClientOriginalExtension();
+            $banner->banner_image = $filename;
+            $path = public_path().'/home/image';
+            $video->move($path, $filename);
+        }
+        else{
+            $banner->banner_image = saveImage([$request->banner_image],100,'public/home/image');
+        }
 		$banner->save();
 		return back();
 	}
